return with 1 if exception is raise, fail test if not
Until now the exception test was done by only raising the exception
and in case of an exception the runtime exited with exit code 1
to make the tests work. This induced the problem that NOT raising
the exception would never be detected. To fix this raising an exception
now lets the program return with exit code 1 (which is the correct
behaviour anyway). To make the test work the exception test is now
executed in a separate rule that requires the exit code to be greater
than zero and that will fail if no exception is raised.